Fix crash in test runs.
Apparently there's a weird circular ref bug in the test logic when using
the command loader which only appears in the tests. By forcing a require
of the commandline module before running the tests the circular refs are
resolved. Very strange, but don't have time to debug further right now.