Refactoring storage of api token
Using ProfileImpl class to centralize logic.

Added methods/accessors from profile instead of standalone methods.

Storing access token id as well as token in the token store as data, so
that we can easily pull it back when needed and not have to break a lot
of existing API surface.