Fix getting crash groups
At the moment, we are sending a `top` parameter of the value `Number.MAX_SAFE_INTEGER`. Unfortunately this number is 9007199254740991 - which is way larger than the value accepted by the backend data store used by the symbol indexing service.

I’ve changed it to use a 32 bit integer to fix this issue.