Add first functional test that uses an account to perform real actions (#718)
Update the Azure Pipelines config to download the API token for an account in the integration environment from a key vault. It passes that token to the PowerShell script which invokes Pester. That script logs in the user, and then tests run using that account. A single simple test creates an app, checks it worked and deletes it again.