Fixes passing token/env into release task from group/store publish (#1004)
The `distribute stores/groups publish` commands work a bit differently under the hood in that they call `distribute release` with the appropriate `--store` or `--group` parameter. They create a new command in doing so, and hence need to pass on the parameters `--env`, `--token`, and `--disable-telemetry` in order for those to work. We missed this in the initial implementation. Other global parameters do not need to be passed, since they set global variables for the life of the call to `appcenter`.