Fix merge-junit-xml and merge-nunit-xml issue (#2235)
* Fix merge-nunit-xml issue

Using the option  --merged-file-name "output.xml" --test-output-dir JunitTestResult, facing the below issue, i can see .zip was was downloaded but merged files is not working!

Downloaded artifacts to JunitTestResult\junit_xml_zip.zip
Error: The "data" argument must be of type string or an instance of Buffer, TypedArray, or DataView. Received an instance of Document

Similer issue with nunit for xamarin uitest: https://github.com/microsoft/appcenter-cli/pull/1814

* Fix merged-file-name issue 

Using the option merged-file-name "output.xml" --test-output-dir JunitTestResult, facing the below issue, i can see .zip was was downloaded but merged files is not working!

Downloaded artifacts to JunitTestResult\junit_xml_zip.zip

Error: The "data" argument must be of type string or an instance of Buffer, TypedArray, or DataView. Received an instance of Document

* Fixed error Delete ···· prettier/prettier

/home/vsts/work/1/s/src/commands/test/run/appium.ts
62:5 error Delete ···· prettier/prettier
63:1 error Delete ···· prettier/prettier

Issue was with file format, fix it!